Alter the way generating file name.
