+++ src/common/utils/index.ts
@@ -14,6 +14,7 @@ export * from './removeEmpty'
 export * from './xss'
 export * from './makeStreamToBuffer'
 export * from './content'
+export * from './randomString'
 
 /**
  * Make a valid user name based on a given email address. It removes all special characters including _.
@@ -52,13 +53,4 @@ export const extractAssetDataFromHtml = (
       }
     })
     .get()
-}
-
-export const randomString = (length: number) => {
-  const chars = '0123456789abcdefghijklmnopqrstuvwxyzABCDEFGHIJKLMNOPQRSTUVWXYZ'
-  let result = ''
-  for (let i = length; i > 0; --i) {
-    result += chars[Math.floor(Math.random() * chars.length)]
-  }
-  return result
 }
